+++ cratedb_toolkit/io/mongodb/extract.py
@@ -69,13 +69,18 @@ import typing as t
 import bson
 from pymongo.collection import Collection
 from rich import progress
+from rich.console import Console
+
+console = Console(stderr=True)
+rich = console
 
 progressbar = progress.Progress(
-    progress.TextColumn("{task.description} ", justify="left"),
+    progress.TextColumn("Processing collection: {task.description} ", justify="left"),
     progress.BarColumn(bar_width=None),
     "[progress.percentage]{task.percentage:>3.1f}% ({task.completed}/{task.total})",
     "•",
     progress.TimeRemainingColumn(),
+    console=console,
 )